// packages/batchprompt/src/types.d.ts
export interface PipelineItem {
    row: Record<string, any>;
    workspace: Record<string, any>;
    stepHistory: Record<string, any>[];
    history: any[];
    originalIndex: number;
    variationIndex?: number;
    accumulatedContent: OpenAI.Chat.Completions.ChatCompletionContentPart[];
}